refactor(doctor): extract discord id repair
